BUG: Fix handling logger redeclaration
Previously, the code would remove the first existing handler, which
removed the FileHandler and left the StreamHandler in place. Then
it would add a new StreamHandler. This meant all output was sent to
console twice and not to any file.

Now we remove all existing handlers, and manually add all handlers
we want to have (without relying on `logging.basicConfig()` to
set up the FileHandler). This means that repeated calls to
setup_logging will establish the correct set up - write to console
and one file with the given path.